grafana webapp: correct mixed-security http/https links
The config.js file used by our snapshot of grafana included in it a
hard-coded http: URL template, because that seemed to be the right
thing to do.  (pmwebd doesn't speak https.)  But for proxied
connections where the webapp is served via https:, the config.js is
wrong.  Fix it to fetch it from window.location.protocol instead.