Avoid recreating the checklist elements on each update
Originally, the code would create a string for all the elements in the
checklist.  This string would be recreated and coverted to HTML on
each update.  Now the list is only created once with tags to be
updated with the values on each update.  This should make it possible
to add graphs that show recent history for the various checklist
items.  The previous approach would refetch the various URL embedded
in the checklist.