Avoid hardcoding the entire vector embedded vector URL for each checklist item
Fully specifying the URL as a static string for each checklist item
will present problems when one wants to get embedded graphs for other
hosts.  The URL has been split into three parts vector root, list of
items to graph, and vector host.  The vector root and vector host are
no longer in the checklist item node initialization string.  Only the
list of items to graph in is in the checklist item node now.  This
should make it easier for the code to get data from other locations.