Add a couple of nodes to differentiate the cause of swapping
Even when a system has a significant amount of memory there are cases
where swapping occurs because a NUMA node is running low on memory.
Thus, there are two nodes in checklist for low memory, one for the
entire system running low on memory and the other for individual NUMA
nodes running low on memory.