Split network node into network transmit and receive nodes
There are separate pcp metrics for network transmit and receive.  It
is a bit easier to keep the metrics separate for those and have
individual nodes for each.  The transmit node is kept as the "Network"
node for the moment because there are other nodes tied to that.
However, once those children nodes are split the network transmit node
node will be relabeld "Network TX" to mirror the "Network RX" node.