Improve metrics and predicate for "Storage - small blocks"
The small blocks node was using the wrong metrics in the computation.
The computations are closer to what is desired.  However, it still
needs some revison.  If there are no disk operations the predicate
ends up being true and when viewing the graphs the derived metric
showing the average size of the io operation is sometimes not plotted.