Add predicate for disk bandwidth use
This is an imperfect predicate for disk use.  The bandwidth value is
hardcoded at 2.5GB/s.  The actual bandwidth availabe is going to
depend on the storage device and the interface that connects it to the
computer.  Also this is probably going to wrong for storage devices
aggregated together as RAID.