Limit graphs to displying non-negative values
Most all of the values that are being plotted are either counters or
changes of counters over times.  There are cases that the counts are
zero but they will not drop below zero. When the values are zero auto
scaling makes the assumption that some values might later be less than
zero and scales the graph to allow the display of negative values.
This patch avoids this display oddity.